Move credential policy out of kobo-net
