+++ crates/kobo-sim/src/lib.rs
@@ -2111,7 +2111,7 @@ fn simulated_tasks(name: &str) -> TaskRunner {
         .with_fetch(Arc::new(kobo_net::fetch_from))
         .with_post(Arc::new(kobo_net::post))
         .with_credential_policy(Arc::new(move |credential, url, usage| {
-            kobo_net::credential_allowed_for(&app, credential, url, usage)
+            kobo_policy::credentials::allowed(&app, credential, url, usage)
         }))
         .with_capabilities([kobo_policy::Capability::Network])
 }
